[FLINK-4093] Expose metric interfaces
